pyproject.toml
[project]
name = "gamgulmarket"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.29.0",
    "fastapi-mctools>=0.5.0",
    "fastapi>=0.115.0",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=10.4.0",
    "pydantic-settings>=2.5.2",
    "pydantic[email]>=2.9.2",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.12",
    "redis>=5.1.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "supabase>=2.8.1",
    "uvicorn>=0.31.0",
]

[tool.uv]
dev-dependencies = [
    "alembic>=1.13.3",
    "psycopg2-binary>=2.9.9",
    "pytest-asyncio>=0.24.0",
    "pytest>=8.3.3",
    "sqlacodegen==3.0.0rc5",
    "sqlalchemy-utils>=0.41.2",
]

[tool.ruff]
line-length = 130


[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings ="ignore::DeprecationWarning"
